decouple file_manipulations
